Excepcion añadida al crud

+++ src/main/java/cat/copernic/roap/Encargos/serveis/CategoriaService.java
@@ -5,6 +5,7 @@
 package cat.copernic.roap.Encargos.serveis;
 
 import cat.copernic.roap.DAO.CategoriasDAO;
+import cat.copernic.roap.DAO.PrendaDAO;
 import cat.copernic.roap.models.Categorias;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,8 +19,12 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class CategoriaService implements CategoriaServiceInterface {
+
     @Autowired
     private CategoriasDAO categoriasDAO;
+    
+    @Autowired
+    private PrendaDAO prenda;
 
     @Override
     @Transactional(readOnly = true)
@@ -36,6 +41,17 @@ public class CategoriaService implements CategoriaServiceInterface {
     @Override
     @Transactional
     public void eliminarCategoria(Categorias categoria) {
+        int idCategoria = categoria.getId();
+
+        // Verificar si existen prendas asociadas a la categoría
+        boolean existenPrendas = prenda.existenPrendasPorCategoria(idCategoria);
+
+        if (existenPrendas) {
+            throw new IllegalStateException("No se puede eliminar la categoría, primero elimine las prendas asociadas");
+            // También podrías retornar un valor booleano en lugar de lanzar una excepción
+            // return false;
+        }
+
         categoriasDAO.delete(categoria);
     }
